+++ src/twisted/internet/endpoints.py
@@ -1138,109 +1138,6 @@ class SSL4ServerEndpoint:
         )
 
 
-def _redirectfy_https(target_protocol):
-    """
-    Wrap the protocol and if first received data is HTTP redirect with a
-    Host header, redirect to HTTPS
-    """
-
-    protocol_receive = target_protocol.dataReceived
-
-    def get_http_host(data):
-        """
-        Return the value of the Host header found inside `data` HTTP
-        raw request.
-        """
-        parts = data.split(b"\r\n")
-        if len(parts) == 1:
-            return b""
-
-        for part in parts:
-            if not part.lower().startswith(b"host"):
-                continue
-
-            segments = part.split(b":", 1)
-            if len(segments) != 2:
-                return b""
-
-            return segments[1].strip()
-
-        return b""
-
-    def first_ssl_receive(data):
-        """
-        Called when initial connection data is received over SSL.
-
-        It checks if an HTTP request is made over HTTPS port, instead of an
-        SSL handshake.
-        """
-        # We only want to handle the first package.
-        target_protocol.dataReceived = protocol_receive
-
-        host = get_http_host(data)
-
-        if not host:
-            # Just forward the data to the wrapped protocol.
-            return protocol_receive(data)
-
-        target_protocol.transport.write(
-            b"HTTP/1.1 301 Moved Permanently\r\n"
-            b"Location: https://%s\r\n"
-            b"Cache-Control: no-store, no-cache\r\n"
-            b"Connection: close\r\n"
-            b"Content-Length: 8\r\n"
-            b"\r\n"
-            b"redirect" % (host,)
-        )
-        target_protocol.transport.loseConnection()
-        return b""
-
-    target_protocol.dataReceived = first_ssl_receive
-
-
-if TLSMemoryBIOFactory:
-
-    class HTTPSMemoryBIOFactory(TLSMemoryBIOFactory):
-        def buildProtocol(self, addr):
-            protocol = self.protocol(self, self.wrappedFactory.buildProtocol(addr))
-            _redirectfy_https(protocol)
-            return protocol
-
-
-class HTTPSServerEndpoint(SSL4ServerEndpoint):
-    """
-    HTTPS secured TCP server endpoint with an IPv4 configuration that
-    redirect HTTP request to HTTPS.
-    """
-
-    def listen(self, protocolFactory):
-        """
-        Implement L{IStreamServerEndpoint.listen} to listen for HTTPS on a
-        TCP socket.
-        """
-        if TLSMemoryBIOFactory is None:
-            return defer.fail(
-                NotImplementedError(
-                    "OpenSSL not available. Try `pip install twisted[tls]`."
-                )
-            )
-
-        def listen_https():
-            """
-            Wrapper for catching all errors into a deferred.
-            """
-            tlsFactory = HTTPSMemoryBIOFactory(
-                self._sslContextFactory, False, protocolFactory
-            )
-            port = self._reactor.listenTCP(
-                self._port, tlsFactory, backlog=self._backlog, interface=self._interface
-            )
-            port._type = "TLS"
-            return port
-
-        return defer.execute(listen_https)
-
-
 @implementer(interfaces.IStreamClientEndpoint)
 class SSL4ClientEndpoint:
     """
